Add files via upload
This script will create a tcp server on a given port and wait to respond to ping requests on that port.
If the request matches a custom string given at the run time it will send a specific response. 
If the ping request does not match exactly it will send a series of different messages to help the user get the right response.